+++ lib/registryFromSparse.nix
@@ -1,4 +1,4 @@
-{ registryFromDownloadUrl, lib }:
+{ registryFromDownloadUrl, lib, pkgsBuildBuild }:
 
 {
   indexUrl,
@@ -7,11 +7,13 @@
 }:
 
 let
+  inherit (pkgsBuildBuild) fetchurl;
+
   slashTerminatedIndexUrl = if lib.hasSuffix "/" indexUrl then indexUrl else "${indexUrl}/";
-  configContents = builtins.readFile "${builtins.fetchurl {
+  configContents = builtins.readFile "${fetchurl (fetchurlExtraArgs // {
     url = "${slashTerminatedIndexUrl}config.json";
     sha256 = configSha256;
-  }}";
+  })}";
 
   config = builtins.fromJSON configContents;
   dl =